+++ modelsTimescale/annotations/annotation.js
@@ -22,11 +22,24 @@ module.exports = (sequelize, DataTypes) => {
     },
     frequency_min: {
       type: DataTypes.INTEGER,
-      allowNull: false
+      allowNull: true
     },
     frequency_max: {
       type: DataTypes.INTEGER,
-      allowNull: false
+      allowNull: true
+    },
+    classification_id: {
+      type: DataTypes.INTEGER
+    },
+    is_manual: {
+      type: DataTypes.BOOLEAN,
+      allowNull: false,
+      defaultValue: true
+    },
+    is_positive: {
+      type: DataTypes.BOOLEAN,
+      allowNull: false,
+      defaultValue: true
     },
     created_by_id: {
       type: DataTypes.INTEGER,