+++ services/detections/index.js
@@ -29,7 +29,6 @@ async function defaultQueryOptions (start, end, streamId, streamsOnlyCreatedBy, 
 
   // TODO: if minConfidence is undefined, get it from event strategy
   condition.confidence = { [models.Sequelize.Op.gte]: (minConfidence !== undefined ? minConfidence : 0.95) }
-
   return {
     where: condition,
     include: [
@@ -54,22 +53,8 @@ async function defaultQueryOptions (start, end, streamId, streamsOnlyCreatedBy, 
   }
 }
 
-async function query (start, end, streamId, classifications, minConfidence, reviews, limit, offset, user) {
+async function query (start, end, streamId, classifications, minConfidence, limit, offset, user) {
   const opts = await defaultQueryOptions(start, end, streamId, undefined, false, classifications, minConfidence, false, limit, offset, user)
-  if (reviews) {
-    opts.include.push({
-      as: 'reviews',
-      model: models.DetectionReview,
-      include: [
-        {
-          as: 'user',
-          model: models.User,
-          attributes: models.User.attributes.lite
-        }
-      ],
-      attributes: ['positive', 'created_at']
-    })
-  }
   return models.Detection.findAll(opts)
 }
 
@@ -100,28 +85,32 @@ function create (detections) {
     })))
 }
 
-function get (detectionId) {
-  return models.Detection.findByPk(detectionId, {
-    include: [
-      {
-        as: 'classification',
-        model: models.Classification,
-        attributes: models.Classification.attributes.lite,
-        required: true
-      },
-      {
-        as: 'stream',
-        model: models.Stream,
-        attributes: models.Stream.attributes.lite
-      },
-      {
-        as: 'classifier',
-        model: models.Classifier,
-        attributes: models.Classifier.attributes.lite
-      }
-    ],
-    attributes: models.Detection.attributes.full
+function matchDetectionsWithReviews (detections, reviews) {
+  detections = detections.map(d => {
+    return {
+      ...d.toJSON(),
+      reviews: []
+    }
   })
+  reviews.forEach((r) => {
+    r = r.toJSON()
+    const detection = detections.find((d) => {
+      return r.start.valueOf() === d.start.valueOf() && r.end.valueOf() === d.end.valueOf() &&
+        r.stream_id === d.stream_id && r.classification.value === d.classification.value
+    })
+    if (detection) {
+      detection.reviews.push({
+        positive: r.is_positive,
+        user: r.created_by
+      })
+    }
+  })
+  return detections
 }
 
-module.exports = { query, timeAggregatedQuery, get, create }
+module.exports = {
+  query,
+  timeAggregatedQuery,
+  create,
+  matchDetectionsWithReviews
+}